fix: change file reference of renamed folder
